Add parser BOM warning test
