+++ Cabal/Cabal.cabal
@@ -303,6 +303,7 @@ extra-source-files:
   tests/PackageTests/UniqueIPID/P2/M.hs
   tests/PackageTests/UniqueIPID/P2/my.cabal
   tests/PackageTests/multInst/my.cabal
+  tests/ParserTests/warnings/bom.cabal
   tests/Setup.hs
   tests/hackage/check.sh
   tests/hackage/download.sh
@@ -651,7 +652,25 @@ test-suite parser-tests
     buildable: False
 
   type: exitcode-stdio-1.0
+  hs-source-dirs: tests
   main-is: ParserTests.hs
+  build-depends:
+    base,
+    bytestring,
+    filepath,
+    tasty,
+    tasty-hunit,
+    tasty-quickcheck,
+    Cabal
+  ghc-options: -Wall
+  default-language: Haskell2010
+
+test-suite parser-hackage-tests
+  if !flag(parsec)
+    buildable: False
+
+  type: exitcode-stdio-1.0
+  main-is: ParserHackageTests.hs
 
   hs-source-dirs: tests
   build-depends:
