+++ Cabal/tests/ParserTests/warnings/bom.cabal
@@ -1,0 +1,7 @@
+﻿name: bom
+version: 1
+cabal-version: >= 1.6
+
+library
+    build-depends: base >= 4.9 && <4.10
+    hs-source-dirs: .
